store 사용자 권한 부여 및 권한 적용

+++ controllers/product.py
@@ -1,17 +1,25 @@
-from flask import request, render_template, redirect, url_for
+from flask import request, render_template, redirect, url_for, session
 from werkzeug.utils import secure_filename
 from .blueprint import product
+from .blueprint import user
+from .auth import check_login, redirect_to_longin_form
 from models.product import Product
 from datetime import datetime
 import os
 
+# 상품 등록 페이지 API
 @product.route('/form', methods=['GET'])
 def form():
+    if not session.get('user_id'):
+        return redirect(url_for('product.get_products'))
+    
     return render_template('product_form.html')
 
-# 상품 등록 페이지 API
 @product.route('/regist', methods=['GET', 'POST'])
 def regist():
+    if not session.get('user_id'):
+        return redirect(url_for('product.get_products'))
+    
     # 전달받은 상품 정보
     form_data = request.form
     thumbnail_img = request.files.get('thumbnail_img')
@@ -42,14 +50,20 @@ def get_products():
 # 상품 삭제 API
 @product.route('/<product_id>/delete', methods=['GET'])
 def delete(product_id):
+    if not session.get('user_id'):
+        return redirect(url_for('product.get_products'))
+    
     Product.delete_one(product_id)
-    return "상품이 정상적으로 삭제되었습니다."
+    return redirect(url_for('product.get_products'))
 
 
 
 # 상품 수정 페이지 API
 @product.route('/<product_id>/edit', methods=['GET'])
 def edit(product_id):
+    if not session.get('user_id'):
+        return redirect(url_for('product.get_products'))
+    
     product = Product.find_one(product_id)
     return render_template('product_edit.html', product=product)
 
@@ -58,6 +72,9 @@ def edit(product_id):
 # 상품 수정 API
 @product.route('/<product_id>/update', methods=['GET', 'POST'])
 def update(product_id):
+    if not session.get('user_id'):
+        return redirect(url_for('product.get_products'))
+    
     form_data = request.form
     thumbnail_img = request.files.get('thumbnail_img')
     detail_img = request.files.get('detail_img')
